+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Added
 
 - [`intl-collator`] Added the custom element
+- [`intl-collator-compare`] Added the custom element
 
 ---